requirements.txt
beautifulsoup4==4.7.1
Django==2.2.6
geopy==1.20.0
geotext==0.4.0
gunicorn==19.9.0
jellyfish==0.7.2
joblib==0.13.2
matplotlib==3.0.3
nltk==3.4.5
numpy==1.16.2
oauth2client==4.1.3
pandas==0.24.2
pycountry==19.8.18
pymongo==3.9.0
python-dotenv==0.10.1
regex==2019.4.14
requests==2.21.0
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.1
sklearn==0.0
spacy==2.1.6
statsmodels==0.9.0
textblob==0.15.3
tflearn==0.3.2
whitenoise==4.1.4
uritemplate==3.0.0
urllib3==1.24.2
